gitconfig: Add aliases to un-/ignore files
This commit adds four new aliases which can be used to ignore and
unignore files.

Sometimes there are files which need to be changed during development,
but these changes shouldn't be committed, but the file changes are
appearing in the `git status` output which can be annoying. To add these
files to .gitignore is also not an option because these particular files
are part of the project.

In such a case the new aliases will be useful. They allow to ignore
files just in the local repository, hence they disappear from the
`git status` output and will be ignored by other git commands as well.