sway: use keyboard layout "en" and the variant "mac"
At work I'm using a MacBook, therefore I'm pretty familiar with this
keyboard layout.

What I really like about this layout is that you can write umlauts and
other special characters without dead keys. And the placing of all the
standard keys match with the "qwerty" variant. This reduces confusions
if I'm not on my computer.